getNumLines should produce the correct value now
git-svn-id: http://svn.dmdirc.com/trunk@1553 00569f92-eb28-0410-84fd-f71c24880f43

+++ src/com/dmdirc/WritableFrameContainer.java
@@ -66,9 +66,9 @@ public abstract class WritableFrameContainer extends FrameContainer {
         final String[] splitLines = line.split("\n");
         lines = splitLines.length;
         for (String splitLine : splitLines) {
-            lines += (int) Math.ceil(splitLine.length() / getMaxLineLength());
+            lines += (int) Math.ceil(splitLine.length() / (double) getMaxLineLength());
         }
         return lines;
-    }    
+    }
     
 }
